Support DuckDuckGoSearchAPI and TavilySearchAPI that return compelete contents instead of snippets as default.
